update $user->last_login on OAuth login (#1183)
This updates the last_login timestamp for OAuth logins.  This fixes a bug where the last_login field is not updated for users who choose to login via Google or FB login.

I also changed the name of the method to `doSilexLogin()` to make it more clear that this is a login success handler.