reduce vertical padding/margin on form controls in edit view (#1199)
The intent of reducing vertical padding/margin on edit controls is to increase the usable space in the editor which increases information density for the user.

With this change, the user can see more fields in the entry, as well as see more entries in the compact list on the left side.