Delta updates cannot have array changes at all
It's not just senses and examples that are difficult to deal with; due
to Mongo limitations around push and pop operations, it's actually not
possible to make any array change become a diff update. So we expand the
"new or deleted senses or examples become non-delta updates" rule to
cover all arrays, not just the senses or examples arrays. This means
that adding a new semantic domain, for example, could end up clobbering
someone else's work. But delta updates are still going to cover 90% of
typical use, where only multi-text fields are populated (or single-item
selection fields, like Part of Speech, which can also be delta updates).