Make rendered entry scrollable on mobile (#1200)
Make the "rendered entry" on mobile scrollable with the entry editor. The rendered-entry on desktop remains sticky at the top of the surface.

I use media breakpoints and two classes dc-rendered-on-desktop and dc-rendered-on-mobile with `display:none` to make the rendered entry show properly in different places based upon viewport size.