Remove entry->writeDiff() code path entirely
Thankfully, we don't need the writeDiff() code path, as our MongoMapper
write() function already creates diffs. All we need to do is load the
entry from Mongo, create the new entry without clobbering old values,
and write that new entry, and the MongoMapper can do the rest. So we get
to keep one single write() code path, which is good for reducing
maintenance burden in the future.