Delta updates to minimize edit conflicts (#1204)
When a user edits an entry, the changes are now sent as deltas instead
of full entries, which turn into Mongo updates of specific fields. This
means that edits from different users on the same entry won't clobber
each other.

Limitations: currently, adding a new array item (usually a meaning or
example sentence, but this also applies to arrays like pictures or
semantic domains) will cause a full update instead of a delta update, so
it will clobber other users' work. To reduce the friction caused by
this, adding a new meaning or example will save the entry twice. The
first entry is saved before the new meaning/example is added, in the
hopes that that save might be a delta update. Then the second save will
happen after the meaning is added, which will be a non-delta update.

The same limitation applies to array deletions (removing a meaning or
example sentence), which have to be full updates instead of deltas.

In the future it might be possible to remove this restriction and allow
delta updates that add or remove a meaning or example.