Adjustable update timer (#1224)
This allows changing the poll update timer from the default 32 seconds.
This will cause everyone logged in to the project to make much more
frequent requests to the server, so we don't want to enable it for
everyone. But for teams that have multiple people working on the same
entry at once, this should make their lives easier.