normalize search string to NFC before comparison
Normalize the search string to NFC since all data in LF is normalized to NFC on disk.  This allows for exact match queries to work regardless of form.

Attempt to fix a bug where the default behavior of ignoring diacritics would cause missing search results for complex scripts with combining characters that are not diacritics (e.g. Japanese or Korean)

fixes #1244