reset input value on blur; hide control on filter
This change adds a ng-blur handler that resets the value of the numeric input to the current entry index.  This fixes the input having a sticky invalid state/red box when the user clicks a button or loses focus on the input.

Additionally, we hide the controls when the filter is active, instead of disabling.