add protractor-fail-fast
This causes the E2E test runner to stop running on the first failed test.  This is advantageous for E2E tests since they take a long time to run, and often the first failing test is where the problem lies.

This should save time when running flaky E2E tests.  We should be able to re-run the tests quicker with fail-fast turned on.