e2e test GHA workflow + unit test reporter (#1278)
This PR adds a separate workflow for e2e test runs. The e2e tests could have been included in the existing unit-test workflow, however I chose to split it out into a separate workflow so that it can be reported as a separate line item in the PR checks.

This PR also makes the following minor changes:

adds a unit test reporter to display test results in a PR comment
adds the protractor-fail-fast capability for e2e tests (stop test runner after first failure).
This is advantageous for E2E tests since they take a long time to run, and often the first failing test is where the problem lies.
This should save time when running flaky E2E tests. We should be able to re-run the e2e tests quicker with fail-fast turned on.

Fixes #1276
Fixes #1269