Combine unit and e2e tests into a single GHA workflow + add e2e test reporter (#1281)
A recent pairing session with @longrunningprocess led us to want to try and combine all tests into a single workflow for efficiency and simplicity. Now that I know the status checks can still be separate for a single workflow, it makes sense to combine all tests into a single workflow and have separate reporters / checks.