use editEntryAndScroll() in deleteEntry()
This replaces the low-level setCurrentEntry() method with the higher-level editEntryAndScroll() method that we use to change entry state.  The higher-level method handles URL updating and MRU cache updates among other things.

This fixes a bug where deleting an entry did not update the URL or update the MRU cache, which could result in an error state under certain circumstances.