check MRU entry actually exists
Add a check to ensure the MRU entry actually exists in the list before setting the entryId.

The previous commit ensures that the MRU cache is updated upon deleting an entry, so I'm not 100% sure that this check is needed, but it shouldn't hurt, so I think it's safe to add as a precaution against an invalid state.