chore: move custom base-php files to proper location (#1300)
Move the files referenced by the base-php image to the base-php folder. They used to be part of the app image, but when we moved to a base-php image build, we forgot to move the custom files. This commit puts them in the appropriate place.

Note, I do not believe that it is necessary to rebuild the base-php image as this change does not involve any functional difference to the existing image.

Fixes #1299