Add default Playwright setup
This is the result of running `npm init playwright`, choosing Typescript
and saying "yes" to the GitHub Actions workflow. I then renamed the
workflow to have an .example extension so that GHA won't run it by
default; to enable it, we would just rename it to .yml.

Next commit will take the example Playright setup and tweak it to be an
actual setup for Language Forge.