First Playwright test
This test just looks for the "Language Forge" title on the homepage, so
it's not all that useful as a test. But it allows us to verify that the
web server is up and running, and serves as a way to show how Playwright
can wait for the web server to come up before running tests.