Prevent npm install from taking forever
Now that we've added the `playwright` NPM package, we need to ensure
that its install script (which tries to download browser binaries) does
not run, as it can take many minutes to download them but they aren't
needed in the `app` container. Then we need to ensure that the node-sass
install script *does* get run, as the `npm --ignore-scripts` flag is
global and would otherwise skip node-sass's install script.