Better way to skip Playwright install process
Playwright has an official environment variable to tell it not to
download & install browser binaries. That's a better way to manage the
install process in the Docker images, as it doesn't rely on whitelisting
packages like node-sass. (We would inevitably forget to whitelist some
other package later on, leading to problems).