Run Playwright tests in docker-compose container
This is still a demo of how to do it, rather than a working example. And
one test is still failing, probably because I have my post-login
verification checks wrong. So this still needs some work. But some tests
are running and passing, so at least the config is correct.