In e2e container, login as test_runner_admin
The test that was failing was the one logging in as admin, which is not
the correct name for the admin user inside the e2e container. So now
that we have Playwright running correctly in the e2e container, we don't
need that test anymore. The `loginAs` function works, and that's what we
want for future tests.