Retry failed tests once and capture trace
This allows for replaying failed tests using `npx playwright show-trace`
and seeing the loaded page, any network traffic, and all sorts of useful
things.

Note that traces are saved in the test-results folder *inside the
container*. So to get them out you'll need to do something like this:

- docker container ls
- Find the name of the container (should be still running)
- docker container cp CONTAINERNAME:/root/test-results .
- npx playwright show-trace test-results/TESTNAME-retry1/trace.zip

At some point we may automate the process of getting the trace.zip files
out of the container if any tests fail, but for now this manual process
is good enough.