Get project config in a test
This allows me to get the project config for the project the admin and
manager users are currently logged into, which includes the update timer
interval. This is needed because we don't have an API to update parts of
the project settings, only to update the whole thing at once. So we need
to fetch the settings, edit the part we need, and write it back, in
order to set up the test conditions.