Make user 2 submit data one second behind user 1
That way we eliminate the scenario where both edits arrive within
milliseconds of each other and Mongo hasn't finished saving edit 2
before edit 1 arrives. That scenario will still cause one of the edits
to be lost, and there's not much we can do about it. But it happens in a
vanishingly rare number of cases. Most of the time edits are separated
by a lot more than a few milliseconds, so this is a more realistic test.

Note that the test still fails.