Fix the bug
The actual fix is a one-liner, once we know what's causing it. It was
normalization: U+0061 LATIN SMALL LETTER A followed by U+0301 COMBINING
ACUTE ACCENT was in the Mongo database (in my test example), but what's
saved into Mongo was U+00E1 LATIN SMALL LETTER A WITH ACUTE. The diff
algorithm naturally compared those two as different (since it's doing a
basic byte comparison on strings), and therefore a delta update was
being submitted which changed á to á. Mongo was therefore writing that
field's value in the database, accidentally overwriting the edit from
the previous user.

The answer is to normalize the pristineEntry value (the value that came
out of the database) before passing it to the diffing algorithm, so that
it's comparing NFC to NFC.

This one-line change makes the E2E test, which had been reliably
failing, pass reliably instead.