Faster test: Wait for API response, not timeout
Instead of waiting for an arbitrary timeout, let the test run faster by
waiting for a specific response from the LF API.

Note that we can't control *when* the API call runs, as it's on an
interval timer controlled by the lexical entry edit page. So instead, we
set the interval timer to a low value, then block the API call until
we're ready for it to be allowed through. That's almost as good as
calling it ourselves from inside the test.