Fix setting interval timeout
Turns out that we have no API for setting the current project. We only
have API calls for: 1) creating a project and making it the current
project, or 2) joining a Send/Receive project and making it the current
project. And the project ID stored in the session wasn't the project ID
of the test project, so the test project wasn't the one whose poll
update interval was getting set.

Now we click on a project link in the UI during global setup, so that
every storage state we save will include the projectId of the test
project. That way we can ensure that we set the interval timer on the
correct project, and greatly accelerate the process of running the test.