Make globalSetup use correct base URL
Now the baseURL from the playwright config is respected everywhere. The
mistake I had made is that by the time the config is put together and
handed to globalSetup, the global `use` section of the config has been
merged into each individual project, and so we need to access
`projects[0].use` to get at it properly.

Now if there is no baseURL, then localhost:(port) is used, but that can
be overridden by specifying a baseURL.