Link to testConstants.json so Typescript is happy
TypeScript was complaining about testConstants.json not being in the
rootDir folder, which is test/e2e. We want a single source of truth,
and we can't move testConstants.json out of test/app yet since we
want the e2e tests in test/app to continue to work for now. So we
create a symlink in test/e2e, which satisfies Typescript that the
file lives under rootDit but lets us still have one source of truth.