Optimize creation of login storage state files (#1354)
We now skip creating the file if it already exists and has a creation
date less than 1 year old (since that's the default lifetime of the
"Remember Me" cookies that Silex creates). This means that the slowness
of logging in five times through the UI (which takes about 20 seconds on
a relatively fast developer machine) will only happen once per machine.