Add "change password" call to test control API
This function lives in a new file called testControl.ts. In the long
run, this file will grow to include all the necessary API calls needed
to set up tests for E2E: creating test projects, assigning and removing
project membership, and so on. For now it just has one function.