Site traversal playwright tests (#1352)
The following tests have been converted to playwright

Bellows E2E Page Traversal
✓ Explore signup page
✓ Explore forgot password page
✓ Explore login page
✓ Explore change password page
✓ Explore activity page
✓ Explore project page
✓ Explore site admin page
✓ Explore user profile page

* Link to testConstants.json so Typescript is happy

TypeScript was complaining about testConstants.json not being in the
rootDir folder, which is test/e2e. We want a single source of truth,
and we can't move testConstants.json out of test/app yet since we
want the e2e tests in test/app to continue to work for now. So we
create a symlink in test/e2e, which satisfies Typescript that the
file lives under rootDir but lets us still have one source of truth.

* Convert page traversal e2e tests from Protractor to Playwright

Run playwright tests as part of a pull request

* Add "change password" call to test control API

This function lives in a new file called testControl.ts. In the long
run, this file will grow to include all the necessary API calls needed
to set up tests for E2E: creating test projects, assigning and removing
project membership, and so on. For now it just has one function.

* Better fixture setup: save username, password, etc

The adminTab fixtures now save the username, password, etc. from the
testConstants file, so that resetting the password is as simply as doing
`changePassword(memberTab.username, memberTab.password)`.

Co-authored-by: Robin Munn <rmunn@pobox.com>
Co-authored-by: Chris Hirt <chris@hirtfamily.net>