Improve E2E test fixtures
* Tab fixtures: extract data setup into own function

This is in preparation for creating new admin/member/etc. fixtures
that parallel adminTab, memberTab, etc., but which don't create a
new browser context; they will only carry data.

* Add new UserDetails fixtures (faster than UserTab)

If you just need user details like username, password, etc., it's overkill
to request a whole adminTab or memberTab just to access its username and
password. Instead, we create 'admin' and 'member' (and so on) fixtures to
just contain the data. These fixtures are extremely fast to create, speeding
up some tests that didn't need a full browser tab.

* Speed up change password tests by reusing fixtures

In some cases, it can make sense to reuse a browser context fixture
for efficiency's sake (it can take a second or two to create a new
browser context). The "change password" tests are one such case,
because the tests don't actually need a new browser tab for the ones
that simply verify whether a button is disabled.

Note that when beforeEach was changed to beforeAll, we had to remove
the page.close() and context.close() calls from the fixture setup.
Otherwise we'd have closed changePasswordPage over that browser tab
instance and then immediately closed it when the fixture went out of
scope at the end of beforeAll. Which would result in every test
failing due to its browser tab having been closed!

In this commit, we also use the more-efficient member fixtures instead
of memberTab when all we want is the username and password.

* Improve TypeScript typing of test function

With the `test.extend` feature in Playwright, it's better to extend it
once with a large number of fixture definitions, rather than extend it
multiple times with one fixture each. The former results in a much nicer
TypeScript type overlay when you over over the `test` function later.