Fix saving data in custom list fields
It was a little tricky to ensure that custom list fields could save data
while not messing up the ability to save data into multitext fields.
There is a quite extensive comment included in the code about how we
needed to distinguish between the two. I may remove or trim down that
comment in the next commit.