Run lfmerge in separate container (#1315)
* Add new lfmerge container to docker-compose.yml
* Add lfmerge container to Kubernetes deployment
* Add environment variables to LfMerge deployment

Add env vars to Docker-Compose and Kubernetes deployments for the
lfmerge container. These won't do anything until the LfMerge change that
looks for the env vars is released, but there's no harm in having them
in here before that change gets released.

* Use specific lfmerge container version in k8s

This ensures that when we want to bump LfMerge to a new release, we'll
be able to just bump the Kubernetes yml file and re-deploy.