Run lfmerge in separate container
This version of the deployment setup uses a .json file at the root of
the repository to keep track of the lfmerge version we want to use. This
can be overridden by using environment variables, e.g. for manual
deployments. But most of the time when a new lfmerge version is desired,
we'll bump the version number recorded in other-dependencies.json.