Update/Add links under Help/About menu (#1377)
Initiated by user feedback, this PR updates some links under the "Help" menu, now renamed to "About". Summary of changes:

Remove the How To / Tutorials link. This just pointed to the community site under a sub-category called How To / Tutorials. It's not well maintained.
Removed the tour link (we no longer pay for HelpHero - it never worked very well)
Add a link to YouTube videos
Add a link to Known issues and limitations on GH wiki
Rename "What's new in the app" to "What's New" and link to the post that @alex-larkin maintains for release summaries
re ordered the links a bit
added font awesome v4 icons