Run lfmerge in separate container (#1378)
The single source of truth for the LfMerge version that will be used is
the version in docker/lfmerge/Dockerfile. In docker-compose.yml, that
file is used directly to build the lfmerge image. In the Kubernetes
deployments, the version number is extracted from that Dockerfile by
using the `cut` command with a colon delimiter and extracting the second
field, and then using that version number in the `kubectl set image`
commands just like we do for the app, next-app, and next-proxy images.