Make counting of projects better suitable for parallelization
When the projects of a user are counted at the beginning of the test
and then at the end of the test again, one does not want tests/projects
from other spec files interfere with it.
The new function countSpecificProjects takes a string and only
counts projects where the name contains this string.