Create EditorPage with new design approach
When using an instance of the page class (e.g. ProjectSettingsPage)
in a spec file, one had to pass the name of a project for every
single method call, e.g. projectSettingsPage.goto('My Project').
The new approach is to pass the information about the project
directly to the constructor.