Commit test with bug so that others can reproduce it
A comment on one entry is made. Thereafter, the page is
reloaded by Playwright's goto. This causes the page to
not display the comments panel. However, the post button
to post a comment is still visible in Playwright's definition
of being visible.
This problem is not there when adding a page.reload in line 89
of editor-comments.spec.ts