Failed to improve editor comments tests
The changes in this commit aimed at making the tests less
flaky for the browsers Webkit and Firefox. Unfortunately, this
was not possible. Moreover, most tests now fail in Chromium as well.
This code is still committed as it might be a good starting point
for future work on these tests.
For now, the conversion of the editor comments tests is abandoned
as the tests' flakiness is mainly caused by the flakiness of Language
Forge itself.