Reduce lfmerge container size (#1405)
The reduced size may not be immediately visible with this commit, as it
will still have to download a 700 MB layer in the Docker image. But with
the changes I've made to the lfmerge build process, that should be the
last time that Docker image layer has to be downloaded, and subsequent
lfmerge releases should be able to reuse the same layer.