Move lfmerge container to separate pod using pod affinity (#1407)
* Fix two wrong env values in lfmerge container

Two env values in the lfmerge container were being read as an int and a
boolean, when Kubernetes requires them all to be strings. So we quote
them to ensure they will be strings.

* Separate pod for lfmerge, with pod affinity

Move lfmerge container to separate pod, using pod affinity to force it
to always run on the same node as the app container so that they can
share the sendreceive volume.

* Speed up docker-compose build of lfmerge container

The lfmerge container doesn't need any files from the languageforge repo
at all, so let's make its Docker build context as small as possible.
That way it's a nearly-instant build once the image has downloaded.