Fix bug cloning FW8 projects
Found and fixed an LfMerge build issue that was causing the LfMerge
build for FieldWorks 8.2 to be incorrectly pulling in the DLLs from
FieldWorks 8.3. This was causing older projects, such as my
test-rmunn-04 project, to fail to clone correctly.

LfMerge version 2.0.115 has the bugfix; this is the version we should
deploy next time we push a Language Forge release.