Rename environment variable used by LfMerge
LfMerge uses an environment variable called LD_TRUST_TOKEN, but
variables starting with LD_ are stripped out of the environment by sudo
for security reasons (setting a different LD_LIBRARY_PATH could be used
for privilege escalation, for example). So we need to rename that
variable so that LfMerge can continue to use it.

This commit also bumps the LfMerge version to the release where this
variable was renamed.