Fix E2E test issues with duplicate projects
A couple different things were causing duplicate E2E projects to be
created with the same projectCode:

* Duplicate project-settings.spec.ts and projects-settings.spec.ts files
* An `import { Project } from 'projects-settings'` line
* deleteOne() instead of deleteMany() in TestControl.php

The `import` line was causing the project settings file to be
re-executed, including its `test.describe()` calls, so that its
beforeAll() was getting called twice. By moving the definition of that
type into a separate file, we avoid that issue. And by removing the
duplicate test file, we avoid the other place where duplicate projects
with the same project code were being created.

In addition, another test was failing due to a timeout if the LD API
server wasn't running. Added the `ld-api` container as a dependency of
the Playwright E2E test container to address this issue.