refactor activity_list_dto_for_project
This change adds projectId to the signature of this API method and also adds authorization logic.

The method name is also changed slightly.

It should be noted that this API method is not used anywhere in the legacy code base so we can change it freely without impact of current functionality