k8s config updated to ensure isolated node for lfmerge (#1433)
* k8s config updated to ensure isolated node for lfmerge

* Remove PHP code that manages LfMerge process

LfMerge process now lives in a different container, so the PHP code that
tries to manage it is no longer needed. In some cases, it will return
errors that incorrectly say that LfMerge is not installed; in other
cases it's just useless code. We can get rid of the entire function
called "startLfMergeIfNeeded", which in turn lets us get rid of a bit
more no-longer-used code that was only used in that function.

* removed config to keep app and lfmerge together

Co-authored-by: Robin Munn <rmunn@pobox.com>