temporarily avoid dashboard links in project page
Based upon my workshop this morning and observing multiple projects experiencing 500 errors in the dashboard, some consistently, this commit reverts the project page links to go directly to the lexicon app, instead of the dashboard.

This can be reverted back to the dashboard once we have more confidence that dashboard is stable and does not block users from getting to the app.