remove all instances of Website class
This is a WIP removal via find/replace and hand editing of code to remove all traces of the Website class which no longer provides sufficient value to the programmer.  LF is now a single website that does not need to be parameterized via a dynamic class instance.

Chore: Add prettier with configuration (#1527)

* Add prettier with initial config
* Add VSCode prettier extension
prettier everything