Updating LF OptionListItem's to use GUID's for keys, instead of the value.
This was mostly already being done, but one place in the html (when the field is uneditable) and one in the php (when option list items are constructed without a given key).

The LFMerge side is still not working, but there aren't any obvious changes required yet in this repo.