adjust audio input system regex to be more inclusive (#1604)
Per a user's report, their audio input system was not properly detected in LF due to a too-strict regex. This loosens the regex a bit to allow LF to properly identify the field as an audio input system.

With this change, a tag must have "-Zxxx" and end with "audio"

Before this change, a non-standard input system tag like lwl-Zxxx-x-minority-audio is not detected as an audio input system (although FLEx handles it fine). After this change, this input system tag should work.