remove all traces of Website class (PHP) and site switching (frontend/backend) (#1592)
This is an epic refactor to remove all traces of the Website class and related plumbing in PHP, as well as various bits that did site switching in both the front and back end. Lots of find/replace and hand editing of code was involved. LF is now a single website that does not need to be parameterized via a dynamic class instance.

Includes 133 file changes (including deleted files) and 160,000+ deleted lines of code