use GHA concurrency for PR builds and testing (#1615)
When a dev pushes new commits to a PR, it is possible that GHA is still running a current build and test workflow. This change enables a concurrency policy that will cancel stale GHA workflows. This is considered an efficiency improvement as otherwise GHA continues to build and test outdated code.

References:
- https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#concurrency
- https://stackoverflow.com/questions/66335225/how-to-cancel-previous-runs-in-the-pr-when-you-push-new-commitsupdate-the-curre/67939898#67939898