Bugfix/integrate not uploading playwright results (#1636)
I noticed on a failed deployment that E2E test results were not being uploaded. The reason was that I moved the results, but didn't adapt the action. 😞

I didn't notice this, because we expect there to be no files if tests are all green. But, I think it makes sense to always save results, because:
1. It's a bit confusing to only save results sometimes
2. Saving results is free
3. It can be valuable to be able to compare failed results to successful results

Note: I opted to only always save screenshots, not traces. The only reason is that the upload takes quite a long time if we include traces. That's because our results our sadly quite messy/bloated due to how we manage pages.