switch to using mariadb for ld-db
This change addresses an issue seen on Apple Silicon, whereby the most recent Docker version on MacOS is more strict when it encounters images not built for arm64.

Switching to mariadb from mysql fixes the issue as mariadb provides both amd64 and arm64 images.  mysql does not, or the package metadata is corrupt in some way that Docker is not expecting.

MariaDB is understood to be a drop-in replacement for MySQL.