remove ui-dist volume on every build
This can avoid tricky problems when switching branches and having run `make dev` on a different branch.  Essentially, it is possible to build and run the application using `make`, but the ui-dist folder will come from whatever the previous run of `make dev` was, even if it was on an earlier commit or branch.  This change ensures that each time the application is built, we start with a fresh ui-dist volume.

The ui-dist volume is only really used in the `make dev` use case, but we include the volume in the app docker compose service to make it work together with the ui-builder service.  This has the unfortunately side effect that the ui-dist volume is active on the app service even when no ui-builder service is active (for doing a webpack watch)