Delete User Account feature polish (#1706)
Fixes #1704

The main purpose of this PR was to fix the problem of seeing deleted users in the user list in the site admin app. I added a isDeleted flag to the UserModel and filtered out users with this flag.

Smaller changes: 
* remove unused references in UserModel
* remove ui-dist volume on every build

This can avoid tricky problems when switching branches and having run `make dev` on a different branch.  Essentially, it is possible to build and run the application using `make`, but the ui-dist folder will come from whatever the previous run of `make dev` was, even if it was on an earlier commit or branch.  This change ensures that each time the application is built, we start with a fresh ui-dist volume.

The ui-dist volume is only really used in the `make dev` use case, but we include the volume in the app docker compose service to make it work together with the ui-builder service.  This has the unfortunately side effect that the ui-dist volume is active on the app service even when no ui-builder service is active (for doing a webpack watch)

* report all errors to the user
We no longer want to swallow and ignore unknown errors.  We want to know about them so we can fix them.

* Additionally, I decided to change the username of a deleted user to something more human readable instead of the ID string.

* fix logic error in system admin check
This fixes a logic error when checking for the system admin role of the current user.  The code was checking the role of the user to be deleted, instead of the role of the current user to determine if the delete operation could proceed.