After cloning new project, go to project when done (#1810)
The "Clone new project" page was setting a locationChange event handler to cancel the clone-status timer. However, since the clone-status page counts as a different locaiton, the event was firing as soon as the user pressed the "Get Started" button, cancelling the clone-status timer too soon so the "cloning new project" page was never detecting that the clone completed.

By simply removing the locationChange event handler, we can let the clone status proceed. It changes from "Getting initial data" to "syncing" pretty soon, but that's fine. Once the status goes to idle, the existing code already cancels the timer and loads the project page, so there's no need for any further changes.