Fix more Typescript "implicit any" index errors
In a couple places, we do `if foo.hasOwnProperty(bar)` followed by
getting the value of `foo[bar]`, but Typescript warns that "string"
isn't a valid property name of `foo`. It should recognize that the
hasOwnProperty check means it's safe to access `foo[bar]`, but it
doesn't. We can either change those to be `if bar in foo` which
Typescript *does* recognize as a type guard, or else we can just leave
the code as-is and add @ts-expect-error comments. Since `bar in foo` is
not quite identical to `foo.hasOwnProperty(bar)` (`bar in foo` climbs
inheritance hierarchies whereas `.hasOwnProperty` doesn't), I chose to
leave the code as-is rather than subtly change its behavior.