Fix permission issues on copied files
The docker cp command was preserving the UID/GID of the copied files
even though I didn't pass it the `-a` parameter (whose purpose is to
preserve the UID/GID of the copied files). To work around this issue, we
set the file ownership to 33/33 before copying the files into Docker.