Require username and password for mongodb (#1813)
* Use MongoDB auth from env vars if available

If MONGODB_USER and MONGODB_PASS are set, they will be used to
authenticate to the MongoDB server. If they are not set, then a
connection request with no authentication will be sent (as per the
existing behavior). This allows us to deploy this change, then set
MONGODB_USER and MONGODB_PASS later and have those changes picked up
without a redeployment.

Note that a corresponding change to LfMerge will also be needed.

* Add auth env vars for LfMerge, other containers

LfMerge env vars won't be used until we deploy a new build of LfMerge
that looks for them, which will be in a PR on the LfMerge repo.

* Use alpha build of LfMerge that handles auth

LfMerge now has an alpha build that handles auth. Once we've proved that
it works, we'll release a full build of LfMerge and bump this version
number again.

* Add commented-out lines to turn on MongoDB auth

Uncommenting these two lines will enable MongoDB auth on local dev. Do
not do so until you have created the `admin` user or you may end up
locked out of your local MongoDB.

* Uncomment MongoDB auth lines

This was needed to make tests pass in CI

* Pass authSource as a separate env var

This will allow us to change the name of the Mongo database we store our
auth in, if in the future we decide not to go with the default name.

* Add Kubernetes secrets for Mongo auth

---------

Co-authored-by: Christopher Hirt <chris@hirtfamily.net>