Update outdated GHA workflow versions
This updates outdated workflow versions to ones that use Node 20,
silencing the warnings from GHA about workflows using Node 16. Also,
the actions/checkout@v3 workflows are deprecated and must be moved to v4
by the end of 2024.

Finally, in the case of a couple of small third-party workflows, I
switched to SHA references as recommended in GHA workflow best practices.